# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Viral Waitlist & Referral Marketing Platform API
  description: |
    REST API for the Viral Waitlist & Referral Marketing Platform.

    ## Authentication

    All API requests require authentication via:
    - **API Key:** Header `X-API-Key: sk_live_xxxxx` or `Authorization: Bearer sk_live_xxxxx`
    - **JWT Token:** For dashboard users, `Authorization: Bearer <jwt>`

    API Key Format:
    - Test: `sk_test_` (test mode)
    - Live: `sk_live_` (production)

    ## Rate Limiting

    Rate limits by tier (per minute):
    - **Free:** 60 requests
    - **Starter:** 60 requests
    - **Pro:** 180 requests
    - **Enterprise:** 600+ requests (custom)

    Response headers include:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
    - `X-Request-ID`: Unique request identifier for debugging

    ## Error Handling

    All error responses follow a consistent format with both a human-readable message and a machine-readable error code:

    ```json
    {
      "error": "User-friendly error message",
      "code": "ERROR_CODE"
    }
    ```

    Error codes use `UPPER_SNAKE_CASE` format for easy programmatic handling.

    ### Common Error Codes

    - `INVALID_INPUT` - Validation error or malformed request
    - `UNAUTHORIZED` - Missing or invalid authentication
    - `FORBIDDEN` - Insufficient permissions
...

...

...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/campaigns/{campaign_id}/email-templates:
    parameters:
      - $ref: '#/components/parameters/CampaignIdParam'

    post:
      tags:
        - Email Templates
      summary: Create email template
      description: Create a new email template for the campaign
      operationId: createEmailTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmailTemplateRequest'
      responses:
        '201':
          description: Email template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Email Templates
      summary: List email templates
      description: Retrieve all email templates for a campaign
      operationId: listEmailTemplates
      parameters:
        - name: type
          in: query
          description: Filter by template type
          schema:
            type: string
            enum: [verification, welcome, position_update, reward_earned, milestone, custom]
      responses:
        '200':
          description: List of email templates
          content:
            application/json:
              schema:
                type: array
                description: Array of email templates. Returns empty array [] if no templates found, never null.
                items:
                  $ref: '#/components/schemas/EmailTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/campaigns/{campaign_id}/email-templates/{template_id}:
    parameters:
      - $ref: '#/components/parameters/CampaignIdParam'
      - $ref: '#/components/parameters/TemplateIdParam'

    get:
      tags:
        - Email Templates
      summary: Get email template
      description: Retrieve a specific email template
      operationId: getEmailTemplate
      responses:
        '200':
          description: Email template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Email Templates
      summary: Update email template
      description: Update an email template
      operationId: updateEmailTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailTemplateRequest'
      responses:
        '200':
          description: Email template updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Email Templates
      summary: Delete email template
      description: Delete an email template
      operationId: deleteEmailTemplate
      responses:
        '204':
          description: Email template deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/campaigns/{campaign_id}/email-templates/{template_id}/send-test:
    post:
      tags:
        - Email Templates
      summary: Send test email
      description: Send a test email using this template
      operationId: sendTestEmail
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
        - $ref: '#/components/parameters/TemplateIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipient_email
              properties:
                recipient_email:
                  type: string
                  format: email
                  description: Email address to send test to
                test_data:
                  type: object
                  description: Test data for template variables
      responses:
        '200':
          description: Test email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sent_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== ANALYTICS ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication (sk_live_* or sk_test_*)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication for dashboard users
  parameters:
    CampaignIdParam:
      name: campaign_id
      in: path
      required: true
      description: Campaign unique identifier
      schema:
        type: string
        format: uuid
...
    TemplateIdParam:
      name: template_id
      in: path
      required: true
      description: Email template unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              value:
                error: "Invalid request format. Please check your JSON syntax."
                code: "INVALID_INPUT"
            missing_field:
              value:
                error: "Required field is missing"
                code: "INVALID_INPUT"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authorization token is missing or invalid"
            code: "UNAUTHORIZED"
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            campaign:
              value:
                error: "Campaign not found"
                code: "CAMPAIGN_NOT_FOUND"
            user:
              value:
                error: "User not found"
                code: "USER_NOT_FOUND"
...
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum requests per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Please try again later."
            code: "RATE_LIMIT_EXCEEDED"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "An internal server error occurred. Please try again later."
            code: "INTERNAL_ERROR"
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: User-friendly error message
        code:
          type: string
          description: Machine-readable error code in UPPER_SNAKE_CASE format
          example: "INVALID_INPUT"
      example:
        error: "Invalid request format"
        code: "INVALID_INPUT"

    # ==================== CORE MODELS ====================
...
    EmailTemplate:
      type: object
      required:
        - id
        - campaign_id
        - name
        - type
        - subject
        - html_body
        - text_body
      properties:
        id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        type:
          type: string
          enum: [verification, welcome, position_update, reward_earned, milestone, custom]
        subject:
          type: string
          maxLength: 255
        html_body:
          type: string
          description: HTML email body with template variables
        text_body:
          type: string
          description: Plain text email body with template variables
        enabled:
          type: boolean
          default: true
        send_automatically:
          type: boolean
          default: true
        variant_name:
          type: string
        variant_weight:
          type: integer
          default: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateEmailTemplateRequest:
      type: object
      required:
        - name
        - type
        - subject
        - html_body
        - text_body
      properties:
        name:
          type: string
          maxLength: 255
        type:
          type: string
          enum: [verification, welcome, position_update, reward_earned, milestone, custom]
        subject:
          type: string
          maxLength: 255
        html_body:
          type: string
        text_body:
          type: string
        enabled:
          type: boolean
          default: true
        send_automatically:
          type: boolean
          default: true
    UpdateEmailTemplateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        subject:
          type: string
          maxLength: 255
        html_body:
          type: string
        text_body:
          type: string
        enabled:
          type: boolean
        send_automatically:
          type: boolean
...